lib/gateway_core/gateway_device.cpp, create device class for handling devices in the ram and flash
